Spostate la pagina della prenotazione nel componente reservation_admin
